remove log trait and transfer method to dtr service
